requirements.txt
pyyaml
requests
pydpkg
GitPython
